requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
idna==2.10
joblib==1.0.0
numpy==1.19.5
pandas==1.2.1
psycopg2-binary==2.8.6
pymystem3==0.2.0
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.0
six==1.15.0
sklearn==0.0
soupsieve==2.1
SQLAlchemy==1.3.22
threadpoolctl==2.1.0
tqdm==4.56.0
urllib3==1.26.2
